% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_genlogis.R
\name{dgenlog}
\alias{dgenlog}
\title{The Generalized logistic distribution}
\usage{
dgenlog(x, a = sqrt(2/pi), b = 0.5, p = 2, location = 0)
pgenlog(q, a = sqrt(2/pi), b = 0.5, p = 2, location = 0)
qgenlog(k, a = sqrt(2/pi), b = 0.5, p = 2, location = 0)
rgenlog(n, a = sqrt(2/pi), b = 0.5, p = 2, location = 0)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{a, b, p}{parameters >= 0, with restrictions.*}

\item{location}{location parameter}

\item{k}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required}
}
\description{
Density, distribution function, quantile function and random generation a generalized logistic distribution.
}
\details{
The used distribution for this package is given by: \deqn{f(x) = ((a + b*(1+p)*(abs(x-location)^p))*exp(-(x-location)*(a+b*(|x-location|^p)))) / ((exp(-(x-location)*(a + b* (|x-location|^p)))+1)^2)}
 
The \code{qgenlog()} returns values for P(X < x).\cr 

The default values for \code{a, b, p and location} produces a function with mean 0 and variance close to 1.\cr 

*Restrictions:\cr 

If \code{p} equals to 0, \code{b} or \code{a} must be 0 otherwise there is identifiability problem.\cr 

The distribution is not defined for \code{a} and \code{b} equal to 0 simultaneously.\cr
}
\examples{
pgenlog(0.5) 
curve(dgenlog(x), xlim = c(-3,3)) 

}
\keyword{d,}
\keyword{p,}
\keyword{q,}
\keyword{r}
