% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_genlogis2.R
\name{genlog_mle2}
\alias{genlog_mle2}
\title{Optimization for a generalized logistic distribution}
\usage{
genlog_mle2(parameters, data, alpha = 0.05)
}
\arguments{
\item{parameters}{Initial values for the parameters to be optimized over in the following order \code{c(a, b, p, location)},
\code{location} can be omioped and will be set to 0.}

\item{data}{This is the the data to be utilized for the estimation}

\item{alpha}{type 1 error parameter for confidente interval}
}
\value{
Same as \code{optim} with extra:
\code{bounds} Confidence interval for maximum likelihood parameter estimatives with of 1-\code(alpha) confidence
}
\description{
Maximum likehood estimation of parameters for a generalized logistic distribution.
}
\details{
Maximum likehood estimation of parameters for the distribution proposed in this package.\cr
This function is an application of \code{optim} as a particular case needed for this distribution using the method "\code{L-BFGS-B}".  \cr
For more information of the output check \code{help(optim)}.\cr

The used distribution for this package is given by: 
\deqn{f(x) = ((a + b*(1+p)*(abs(x-location)^p))*exp(-(x-location)*(a+b*(|x-location|^p)))) / ((exp(-(x-location)*(a + b* (|x-location|^p)))+1)^2)}

\code{help(dgenlog)} for parameters restrictions.\cr
}
\examples{
datas <- rgenlog(10000, 1.5,2,2, 0)
genlog_mle2(c(.5,1.6, 1.5, 0),datas, alpha = 0.05)

}
\keyword{d,}
\keyword{genlogis}
\keyword{p,}
\keyword{q,}
\keyword{r,}
