% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider_genlogis.R
\name{genlog_slider}
\alias{genlog_slider}
\title{Slider for generalized logistic}
\usage{
genlog_slider(data, return_var = NULL, loc_range = 10)
}
\arguments{
\item{data}{vector of data to compare.}

\item{return_var}{a char string to name where parameters are assigned}

\item{loc_range}{a number to setup the minimum and maximum range value of the location parameter}
}
\value{
The function plots a interactive graphic in RStudio Viewer panel. \cr
Also, the parameters \code{a, b, p} and \code{location} can be returned to 
\code{return_var} if asked in the graphic.
}
\description{
Make a generalized logistic distribution slider to compare histogram with theoretical distribution
}
\details{
There is a small gear in the top left of the graphic where you can slide the parameters \code{@param a,b,p,location}.
The used distribution for this package is given by: \deqn{f(x) = ((a + b*(1+p)*(abs(x-location)^p))*exp(-(x-location)*(a+b*(|x-location|^p)))) / ((exp(-(x-location)*(a + b* (|x-location|^p)))+1)^2)}
If the density function is not printed it is not defined for these parameters. \cr 

\code{help(dgenlog)} for parameters restrictions.\cr 

This function requires RStudio to run.
\cr
}
\examples{
datas <- rgenlog(1000)
genlog_slider(datas, return_var = 'parameters')

}
\references{
RATHIE, P. N., SWAMEE, P. K. \emph{On a new invertible generalized logistic distribution
approximation to normal distribution}, Technical Research Report in Statistics, 07/2006,
Dept. of Statistics, Univ. of Brasilia, Brasilia, Brazil. 2006.
}
\keyword{genlogis}
\keyword{slider,}
