% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_genlogis.R
\name{qgenlog}
\alias{qgenlog}
\title{The Generalized logistic distribution}
\usage{
dgenlog(x, a = 2/pi, b = 0.5, p = 2, location = 0)
pgenlog(q, a = 2/pi, b = 0.5, p = 2, location = 0)
qgenlog(k, a = 2/pi, b = 0.5, p = 2, location = 0)
rgenlog(n, a = 2/pi, b = 0.5, p = 2, location = 0)
}
\arguments{
\item{k}{vector of probabilities.}

\item{a, b, p}{parameters >= 0, with restrictions.}

\item{location}{location parameter}

\item{x, q}{vector of quantiles.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required}
}
\description{
Density, distribution function, quantile function and random generation a generalized logistic distribution.
}
\details{
The used distribution for this package is given by: \deqn{f(x) = ((a + b*(1+p)*(abs(x-location)^p))*exp(-(x-location)*(a+b*(|x-location|^p)))) / ((exp(-(x-location)*(a + b* (|x-location|^p)))+1)^2)}
 
The \code{qgenlog()} returns values for P(X < x).

The default values for \code{@param a,b,p,location} produces a function with mean 0 and variance close to 1.
}
\examples{
pgenlog(0.5) 
curve(dgenlog(x), xlim = c(-3,3)) 

}
\keyword{d,}
\keyword{p,}
\keyword{q,}
\keyword{r}
