% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_genlogis.R
\name{rgenlog}
\alias{rgenlog}
\title{The Generalized logistic distribution}
\usage{
dgenlog(x, a, b, p)
pgenlog(q, a, b, p)
qgenlog(k, a, b, p)
rgenlog(n, a, b, p)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required}

\item{a, b, p}{parameters > 0.}

\item{x, q}{vector of quantiles.}

\item{k}{vector of probabilities.}
}
\description{
Density, distribution function, quantile function and random generation a generalized logistic distribution.
}
\details{
The used distribution for this package is given by: \deqn{((a + b*(1+p)*(abs(x)^p))*exp(-x*(a+b*(|x|^p)))) / ((exp(-x*(a + b* (|x|^p)))+1)^2)}

The \code{qgenlog()} returns values for P(X < x).

The default values for \code{a, b and p} are \code{sqrt(2/pi), 0.5 and 2} which produces a functions with mean 0 and variance close to 1.
}
\examples{

}
\keyword{d,}
\keyword{p,}
\keyword{q,}
\keyword{r}
