% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider_genlogis.R
\name{genlog_slider}
\alias{genlog_slider}
\title{Slider for generalized logistic}
\usage{
genlog_slider(data)
}
\arguments{
\item{data}{vector of data to compare.}
}
\description{
Make a generalized logistic distribution slider to compare histogram with theoretical distribution
}
\details{
There is a small gear in the top left of the graphic where you can slide the parameters \code{@param a,b,p,location}.
The used distribution for this package is given by: \deqn{f(x) = ((a + b*(1+p)*(abs(x-location)^p))*exp(-(x-location)*(a+b*(|x-location|^p)))) / ((exp(-(x-location)*(a + b* (|x-location|^p)))+1)^2)}
If the density function is not printed it is not defined for these parameters. \cr 

\code{help(dgenlog)} for parameters restrictions.\cr 

This function requires RStudio to run.\cr
}
\examples{
genlog_slider(rgenlog(1000))
}
\keyword{genlogis}
\keyword{slider,}
